Working mpl::insert
An implementation of mpl::insert that accumulates everything up to the
iterator passed in from the original sequence, push_frontS on the rest
of the sequence (a.k.a. the iterator that specifies the position) and
concatenates the result. Horribly slow. Maybe some more helpers for
nlist are in order to ease the pain associated with those operations.