Rewrote take.
It now behaves as a proper take, the new insert_impl should use
size_impl to avoid too cut down the number of specializations.  In
general, I should never specialize on nlist if I don't need the
arguments. The nlist_tag dispatch is going to prevent instantiation
with bad arguments.